Break out util functions
